Change Website Title, Increase Name Size, Adjust Subheading Placement
Title was changed from React App to page specific titles, such as About - PeytonSchaefer for the About page.  My name in the header was increased in size.  The horizontal line was changed from the full width of the page to only 95%.  Finally, the subheading was centered and a smalled horizontal line was added below it.